# src/swagger/swagger.yaml
...
openapi: '3.0.0'
info:
  title: 'Healthcare Interoperability: Organization Ops'
  version: '1.0.0'
  description: 'Interoperability in health care service:  Organization usecase'
servers:
  - url: http://localhost:3000/api/v1
    description: Dev Server
  - url: http://localhost/api/v1
    description: Docker 
  - url: http://192.168.8.190/api/v1
    description: raspberry pi 
paths:
...
  /organization:
    post:
      tags: 
        - Organization
      description: Create an Organization
      requestBody:
        required: 
          - name
          - password
          - telecom
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Parklane Teaching Hospital Enugu
                password: 
                  type: string
                  example: password1234
                type: 
                  type: object
                  example: 
                    system: 'www.system.com'
                    code: 342334343
                    display: University teaching hospital
                identifier:
                  type: array
                  example:
                    - use: 'Official use'
                      system: 'www.system.com'
                      value: Parklane Teaching Hospital Enugu
                telecom:
                  type: array
                  example:
                    - system: email
                      rank: 1
                      value: parklanhospital.gmail.com
                      use: work
                active: 
                  type: boolean
                  example: 
                    true
      responses:
        201:
          description: Created Successfully
        500:
          description: Internal Server Error
    
    get:
      tags: 
        - Organization
      description: Get all the organization
      security:
        - bearerAuth: []
      responses: 
        200: 
          description: Successfull
        401: 
          description: Unathaurized Access
        500: 
          description: Internal Server Error
    
    patch:
      tags:
        - Organization
      description: Update the current logged in organization details
      security: 
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: 
                name: 
                  type: string
                  example: Patrick Otiji specialist hospital updated again
      responses:
        200:
          description: Successfull
        401:
          description: Unauthorized Access
        500: 
          description: Internal Server Error
...
  /contact:
    post:
      tags:
        - Contact
      description: Create a contact person for this organization
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                purpose:
                  type: string
                  example: BILL
                name:
                  type: object
                  example:
                    family: Uzoagulu 
                    given:  Joshua Chukwudike
                    prefix: Mr
                    suffix: OON
                telecom:
                  type: array
                  example:
                    - use: officail
                      rank: 1
                      system: phone
                      value: 08034915305
      responses:
        201:
          description: Created successfully
        500: 
          description: internal server error
        403:
          description: you do not have permission to perform this action
        401:
          description: unauthorized access
...
  /_personnel:
    post:
      tags:
        - _Personnel
      description: Create Personnel for current logged in organization
      security:
        - bearerAuth: []
      responses:
        201:
          description: Successfull
        401: 
          description: Unathurized Access
        403:
          description: You do not have permission to perform this action
        500:
          description: Internal Server Error
    get:
      tags:
        - _Personnel
      description: Get _Personnel of the current logged in organization
      security:
        - bearerAuth: []
      responses:
        201:
          description: Successfull
        401: 
          description: Unathurized Access
        403:
          description: You do not have permission to perform this action
        500:
          description: Internal Server Error
...
  /practitioner:
    post: 
      tags:
        - Practitioner
      description: Add a Practitioner to your organization
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: object
                  example:
                    family: Careful
                    given:
                      - Adam
                      - Luckster
                    prefix:
                      - Dr
                gender:
                  type: string
                  example: Male
                active:
                  type: boolean
                  example: true
                address:
                  type: array
                  example:
                    - use: Official
                      postalcode: 400221
                      line: 
                        - Apo bridge
                        - Asokoro
                birthDate:
                  type: string,
                  example: 12/10/1979
      responses:
        200:
          description: Successfull
        401: 
          description: Unauthorized Access
        500:
          description: Internal server error
    
    get:
      tags:
        - Practitioner
      description: Get all practitioners of this organization
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successfull
        401: 
          description: Unauthorized Access
        500:
          description: Internal server error
...
  /_service:
    post:
      tags: 
        - _Services
      description: Create _service for logged in organization
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                outpatientService:
                  type: boolean
                  example: true
                inpatientServices:
                  type: boolean
                  example: true
                surgicalServices:
                  type: array
                  example:
                    - appendix
                obstericsAndGynecologyServices:
                  type: array
                  example: 
                    - CS
      responses:
        204: 
          description: null
        500: 
          description: internal server error
        403:
          description: you do not have permission to perform this action
        401:
          description: unauthorized access
          
    get:
      tags:
        - _Services
      description: Get _Service of the current logged in user
      security:
        - BasicAuth: []
      responses:
        204: 
          description: null
        500: 
          description: internal server error
        403:
          description: you do not have permission to perform this action
        401:
          description: unauthorized access

    patch:
      tags:
        - _Services
      description: update the _Service of the current logged in organization
      security:
        - bearerAuth: []
      requestBody:  
        content: 
          application/json:
            schema:
              type: object
              properties:
                pediatricsServices:
                  type: array
                  example: 
                    - Crawling
                dentalServices:
                  type: array
                  example:
                    - Washing
                specialClinicalServices:
                  type: array
                  example:
                    - optical
                ambulanceServices:
                  type: boolean
                  example: false
      responses:
        204: 
          description: null
        500: 
          description: internal server error
        403:
          description: you do not have permission to perform this action
        401:
          description: unauthorized access
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
...